Move election lookup widget to module.
